+++ examples/ts/src/constant.ts
@@ -1,1 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
 export const DEFAULT_TOKEN_DECIMAL = 18;