+++ examples/ts/src/md5.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/prefer-default-export */
 import crypto from 'crypto';
 
 export const md5 = (x: $TSFixMe) => crypto.createHash('md5').update(x).digest('hex');