+++ packages/eslint-config-ts/index.js
@@ -1,8 +1,18 @@
 // follow javascript style from [here](https://github.com/airbnb/javascript)
 
-const shared = require("./shared");
+const shared = require('./shared');
 
 module.exports = {
   ...shared,
-  extends: ["@arcblock/eslint-config", "airbnb-typescript"],
+  extends: ['@arcblock/eslint-config', 'airbnb-typescript'],
+  settings: {
+    'import/resolver': {
+      node: {
+        extensions: ['.js', '.jsx', '.mjs', '.ts', '.tsx'],
+      },
+    },
+  },
+  rules: {
+    'require-await': 'error',
+  },
 };